// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plan = {
  FREE: 'FREE',
  PRO: 'PRO'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const ItemContentMode = {
  TEXT: 'TEXT',
  FILE: 'FILE',
  URL: 'URL'
} as const

export type ItemContentMode = (typeof ItemContentMode)[keyof typeof ItemContentMode]